requirements.txt
fastapi
uvicorn[standard]
python-multipart
ollama
python-docx
pymupdf
pytest
pytest-asyncio
httpx
